+++ src/main/java/com/ohgiraffers/geogieoddae/restaurant/command/repository/restaurant/RestaurantPictureRepository.java
@@ -1,4 +1,4 @@
-package com.ohgiraffers.geogieoddae.restaurant.command.repository;
+package com.ohgiraffers.geogieoddae.restaurant.command.repository.restaurant;
 
 import com.ohgiraffers.geogieoddae.restaurant.command.entity.restaurant.RestaurantPictureEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
