+++ src/main/java/com/ohgiraffers/geogieoddae/restaurant/command/repository/restaurant/RestaurantRepository.java
@@ -1,4 +1,4 @@
-package com.ohgiraffers.geogieoddae.restaurant.command.repository;
+package com.ohgiraffers.geogieoddae.restaurant.command.repository.restaurant;
 
 import com.ohgiraffers.geogieoddae.restaurant.command.entity.restaurant.RestaurantEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
