+++ src/main/java/com/ohgiraffers/geogieoddae/restaurant/command/repository/review/ReviewRepository.java
@@ -1,4 +1,4 @@
-package com.ohgiraffers.geogieoddae.restaurant.command.repository;
+package com.ohgiraffers.geogieoddae.restaurant.command.repository.review;
 
 import com.ohgiraffers.geogieoddae.restaurant.command.entity.review.ReviewEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
